pyproject.toml
[project]
name = "hello-world"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipython>=9.7.0",
    "langchain>=1.0.7",
    "langchain-google-genai>=3.1.0",
    "langgraph>=1.0.3",
    "python-dotenv>=1.2.1",
]
